// src/api-client/models/AdminCreateInstaller200Response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import type { Installer } from "./Installer";
import {
  InstallerFromJSON,
  InstallerFromJSONTyped,
  InstallerToJSON,
} from "./Installer";
import type { ResponseMeta } from "./ResponseMeta";
import {
  ResponseMetaFromJSON,
  ResponseMetaFromJSONTyped,
  ResponseMetaToJSON,
} from "./ResponseMeta";

/**
 *
 * @export
 * @interface AdminCreateInstaller200Response
 */
export interface AdminCreateInstaller200Response {
  /**
   *
   * @type {ResponseMeta}
   * @memberof AdminCreateInstaller200Response
   */
  meta: ResponseMeta;
  /**
   *
   * @type {Installer}
   * @memberof AdminCreateInstaller200Response
   */
  result: Installer;
}

/**
 * Check if a given object implements the AdminCreateInstaller200Response interface.
 */
export function instanceOfAdminCreateInstaller200Response(
  value: object,
): boolean {
  let isInstance = true;
  isInstance = isInstance && "meta" in value;
  isInstance = isInstance && "result" in value;

  return isInstance;
}

export function AdminCreateInstaller200ResponseFromJSON(
  json: any,
): AdminCreateInstaller200Response {
  return AdminCreateInstaller200ResponseFromJSONTyped(json, false);
}

export function AdminCreateInstaller200ResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): AdminCreateInstaller200Response {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    meta: ResponseMetaFromJSON(json["meta"]),
    result: InstallerFromJSON(json["result"]),
  };
}

export function AdminCreateInstaller200ResponseToJSON(
  value?: AdminCreateInstaller200Response | null,
): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    meta: ResponseMetaToJSON(value.meta),
    result: InstallerToJSON(value.result),
  };
}