// src/api-client/models/Installer.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Quatt API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.2
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
/**
 *
 * @export
 * @interface Installer
 */
export interface Installer {
  /**
   *
   * @type {string}
   * @memberof Installer
   */
  id: string;
  /**
   *
   * @type {string}
   * @memberof Installer
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof Installer
   */
  code: string;
  /**
   *
   * @type {string}
   * @memberof Installer
   */
  phone: string;
  /**
   *
   * @type {boolean}
   * @memberof Installer
   */
  isActive: boolean;
  /**
   *
   * @type {Date}
   * @memberof Installer
   */
  createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof Installer
   */
  updatedAt: Date;
}

/**
 * Check if a given object implements the Installer interface.
 */
export function instanceOfInstaller(value: object): boolean {
  let isInstance = true;
  isInstance = isInstance && "id" in value;
  isInstance = isInstance && "name" in value;
  isInstance = isInstance && "code" in value;
  isInstance = isInstance && "phone" in value;
  isInstance = isInstance && "isActive" in value;
  isInstance = isInstance && "createdAt" in value;
  isInstance = isInstance && "updatedAt" in value;

  return isInstance;
}

export function InstallerFromJSON(json: any): Installer {
  return InstallerFromJSONTyped(json, false);
}

export function InstallerFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): Installer {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    id: json["id"],
    name: json["name"],
    code: json["code"],
    phone: json["phone"],
    isActive: json["isActive"],
    createdAt: new Date(json["createdAt"]),
    updatedAt: new Date(json["updatedAt"]),
  };
}

export function InstallerToJSON(value?: Installer | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    id: value.id,
    name: value.name,
    code: value.code,
    phone: value.phone,
    isActive: value.isActive,
    createdAt: value.createdAt.toISOString(),
    updatedAt: value.updatedAt.toISOString(),
  };
}